CartApi/pyproject.toml
[project]
requires-python = ">=3.13.5"
name = "cart-api"
version = "0.0.1"
description = "A Restful api for Groceries and Grocery lists"
readme = { file = "README.md", content-type = "text/markdown" }
dependencies = [
    "django",
    "djangorestframework",
    "ruff"
]

[project.optional-dependencies]
linter = ["ruff"]

[tool.pylint.main]
ignore = [
    "**/migrations/**"
]

[tool.ruff]
line-length = 88
indent-width = 4
exclude = [
    "**/migrations/**"
]

[tool.ruff.lint]
extend-select = ["E501"]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "I",
    "PL",
    "FAST",
]

[tool.ruff.lint.isort]
case-sensitive = false
force-single-line = false
lines-after-imports = 2
lines-between-types = 1
no-sections = false
relative-imports-order = "furthest-to-closest"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true